improve error handling and logging
Sometimes under load, the internal timer in the server would
executed substantially later than the pipe reads/writes and
so the deadlines from the read/writes would end first and
this error was not handled and returned an 500 error.
This was changed so that it now properly returns an 416 error.

Additionally, for some error logging, the input and timeout is
also given.